Move material textures loading to own file
