+++ src/ts/frontend/assets/procedural/solarPanel/solarPanelMaterial.ts
@@ -19,9 +19,9 @@ import { NodeMaterialModes } from "@babylonjs/core/Materials/Node/Enums/nodeMate
 import { NodeMaterial } from "@babylonjs/core/Materials/Node/nodeMaterial";
 import { Scene } from "@babylonjs/core/scene";
 
-import * as BSL from "@/utils/bsl";
+import { PBRTextures } from "@/frontend/assets/textures/materials";
 
-import { PBRTextures } from "../../textures";
+import * as BSL from "@/utils/bsl";
 
 export class SolarPanelMaterial extends NodeMaterial {
     constructor(textures: Omit<PBRTextures, "ambientOcclusion">, scene: Scene) {
