+++ src/ts/frontend/assets/procedural/spaceStation/climber/spaceElevatorClimber.ts
@@ -22,7 +22,7 @@ import { CreateBox, CreateTube, TransformNode } from "@babylonjs/core/Meshes";
 import { Mesh } from "@babylonjs/core/Meshes/mesh";
 import { Scene } from "@babylonjs/core/scene";
 
-import { PBRTextures } from "@/frontend/assets/textures";
+import { PBRTextures } from "@/frontend/assets/textures/materials";
 import { ObjectTargetCursorType, Targetable, TargetInfo } from "@/frontend/universe/architecture/targetable";
 
 import i18n from "@/i18n";
