+++ src/ts/frontend/assets/procedural/spaceStation/metalSectionMaterial.ts
@@ -19,9 +19,9 @@ import { NodeMaterialModes } from "@babylonjs/core/Materials/Node/Enums/nodeMate
 import { NodeMaterial } from "@babylonjs/core/Materials/Node/nodeMaterial";
 import { Scene } from "@babylonjs/core/scene";
 
-import * as BSL from "@/utils/bsl";
+import { PBRTextures } from "@/frontend/assets/textures/materials";
 
-import { PBRTextures } from "../../textures";
+import * as BSL from "@/utils/bsl";
 
 export class MetalSectionMaterial extends NodeMaterial {
     constructor(name: string, textures: PBRTextures, scene: Scene) {
