+++ src/ts/frontend/assets/textures/index.ts
@@ -24,28 +24,17 @@ import { Scene } from "@babylonjs/core/scene";
 
 import { ILoadingProgressMonitor } from "../loadingProgressMonitor";
 import { GasPlanetTextures, loadGasPlanetTextures } from "./gasPlanet";
+import { AllMaterialTextures, loadMaterialTextures } from "./materials";
 import { loadRingsTextures, RingsTextures } from "./rings";
 import { createTexturePools, TexturePools } from "./texturePools";
 import { loadCubeTextureAsync, loadTextureAsync } from "./utils";
 
 import butterflyTexture from "@assets/butterfly.webp";
-import crateAlbedo from "@assets/crateMaterial/space-crate1-albedo.webp";
-import crateAmbientOcclusion from "@assets/crateMaterial/space-crate1-ao.webp";
-import crateMetallicRoughness from "@assets/crateMaterial/space-crate1-metallic-roughness.webp";
-import crateNormal from "@assets/crateMaterial/space-crate1-normal-dx.webp";
-import concreteAlbedo from "@assets/degraded-concrete-ue/degraded-concrete_albedo.webp";
-import concreteAmbientOcclusion from "@assets/degraded-concrete-ue/degraded-concrete_ao.webp";
-import concreteMetallicRoughness from "@assets/degraded-concrete-ue/degraded-concrete_metallic_roughness.webp";
-import concreteNormal from "@assets/degraded-concrete-ue/degraded-concrete_normal-dx.webp";
 import flareParticle from "@assets/flare.png";
 import grassAlbedoRoughnessMap from "@assets/grassMaterial/wispy-grass-meadow_albedo_roughness.webp";
 import grassNormalMetallicMap from "@assets/grassMaterial/wispy-grass-meadow_normal_metallic.webp";
 import snowAlbedoRoughnessMap from "@assets/iceMaterial/ice_field_albedo_roughness.webp";
 import snowNormalMetallicMap from "@assets/iceMaterial/ice_field_normal_metallic.webp";
-import metalPanelsMetallicRoughness from "@assets/metalPanelMaterial/metallicRoughness.webp";
-import metalPanelsAlbedo from "@assets/metalPanelMaterial/sci-fi-panel1-albedo.webp";
-import metalPanelsAmbientOcclusion from "@assets/metalPanelMaterial/sci-fi-panel1-ao.webp";
-import metalPanelsNormal from "@assets/metalPanelMaterial/sci-fi-panel1-normal-dx.webp";
 import empty from "@assets/oneBlackPixel.webp";
 import seamlessPerlin from "@assets/perlin.webp";
 import rockAlbedoRoughnessMap from "@assets/rockMaterial/layered-planetary_albedo_roughness.webp";
@@ -53,18 +42,10 @@ import rockNormalMetallicMap from "@assets/rockMaterial/layered-planetary_normal
 import sandAlbedoRoughnessMap from "@assets/sandMaterial/wavy-sand_albedo_roughness.webp";
 import sandNormalMetallicMap from "@assets/sandMaterial/wavy-sand_normal_metallic.webp";
 import skyBox from "@assets/skybox/milkyway.env";
-import solarPanelMetallicRoughness from "@assets/SolarPanelMaterial/metallicRougness.webp";
-import solarPanelAlbedo from "@assets/SolarPanelMaterial/SolarPanel002_2K-PNG_Color.webp";
-import solarPanelNormal from "@assets/SolarPanelMaterial/SolarPanel002_2K-PNG_NormalDX.webp";
-import spaceStationMetallicRoughness from "@assets/spaceStationMaterial/metallicRoughness.webp";
-import spaceStationAlbedo from "@assets/spaceStationMaterial/spaceship-panels1-albedo.webp";
-import spaceStationAmbientOcclusion from "@assets/spaceStationMaterial/spaceship-panels1-ao.webp";
-import spaceStationNormal from "@assets/spaceStationMaterial/spaceship-panels1-normal-dx.webp";
 import cursorImage from "@assets/textures/hoveredCircle.png";
 import plumeParticle from "@assets/textures/plume.png";
 import waterNormal1 from "@assets/textures/waterNormalMap3.jpg";
 import waterNormal2 from "@assets/textures/waterNormalMap4.jpg";
-import treeTexturePath from "@assets/tree/Tree.png";
 
 // Define texture groups types
 export type TerrainTextures = {
@@ -77,13 +58,6 @@ export type AllTerrainTextures = {
     grass: TerrainTextures;
     snow: TerrainTextures;
     sand: TerrainTextures;
-};
-
-export type PBRTextures = {
-    albedo: Texture;
-    normal: Texture;
-    metallicRoughness: Texture;
-    ambientOcclusion: Texture;
 };
 
 export type WaterTextures = {
@@ -99,15 +73,6 @@ export type ParticleTextures = {
 
 export type NoiseTextures = {
     seamlessPerlin: Texture;
-};
-
-export type AllMaterialTextures = {
-    solarPanel: Omit<PBRTextures, "ambientOcclusion">;
-    spaceStation: PBRTextures;
-    metalPanels: PBRTextures;
-    concrete: PBRTextures;
-    crate: PBRTextures;
-    tree: Pick<PBRTextures, "albedo">;
 };
 
 export type Textures = {
@@ -204,96 +169,6 @@ export async function loadTextures(scene: Scene, progressMonitor: ILoadingProgre
     const seamlessPerlinPromise = loadTextureAsync("SeamlessPerlin", seamlessPerlin, scene, progressMonitor);
     const milkyWayPromise = loadCubeTextureAsync("SkyBox", skyBox, scene, progressMonitor);
 
-    // Material textures
-    // Solar Panel
-    const solarPanelAlbedoPromise = loadTextureAsync("SolarPanelAlbedo", solarPanelAlbedo, scene, progressMonitor);
-    const solarPanelNormalPromise = loadTextureAsync("SolarPanelNormal", solarPanelNormal, scene, progressMonitor);
-    const solarPanelMetallicRoughnessPromise = loadTextureAsync(
-        "SolarPanelMetallicRoughness",
-        solarPanelMetallicRoughness,
-        scene,
-        progressMonitor,
-    );
-
-    // Space Station
-    const spaceStationAlbedoPromise = loadTextureAsync(
-        "SpaceStationAlbedo",
-        spaceStationAlbedo,
-        scene,
-        progressMonitor,
-    );
-    const spaceStationNormalPromise = loadTextureAsync(
-        "SpaceStationNormal",
-        spaceStationNormal,
-        scene,
-        progressMonitor,
-    );
-    const spaceStationMetallicRoughnessPromise = loadTextureAsync(
-        "SpaceStationMetallicRoughness",
-        spaceStationMetallicRoughness,
-        scene,
-        progressMonitor,
-    );
-    const spaceStationAmbientOcclusionPromise = loadTextureAsync(
-        "SpaceStationAmbientOcclusion",
-        spaceStationAmbientOcclusion,
-        scene,
-        progressMonitor,
-    );
-
-    // Metal Panels
-    const metalPanelsAlbedoPromise = loadTextureAsync("MetalPanelsAlbedo", metalPanelsAlbedo, scene, progressMonitor);
-    const metalPanelsNormalPromise = loadTextureAsync("MetalPanelsNormal", metalPanelsNormal, scene, progressMonitor);
-    const metalPanelsMetallicRoughnessPromise = loadTextureAsync(
-        "MetalPanelsMetallicRoughness",
-        metalPanelsMetallicRoughness,
-        scene,
-        progressMonitor,
-    );
-    const metalPanelsAmbientOcclusionPromise = loadTextureAsync(
-        "MetalPanelsAmbientOcclusion",
-        metalPanelsAmbientOcclusion,
-        scene,
-        progressMonitor,
-    );
-
-    const treeAlbedoPromise = loadTextureAsync("TreeAlbedo", treeTexturePath, scene, progressMonitor);
-
-    // Concrete
-    const concreteAlbedoPromise = loadTextureAsync("ConcreteAlbedo", concreteAlbedo, scene, progressMonitor);
-    const concreteNormalPromise = loadTextureAsync("ConcreteNormal", concreteNormal, scene, progressMonitor);
-    const concreteMetallicRoughnessPromise = loadTextureAsync(
-        "ConcreteMetallicRoughness",
-        concreteMetallicRoughness,
-        scene,
-        progressMonitor,
-    );
-    const concreteAmbientOcclusionPromise = loadTextureAsync(
-        "ConcreteAmbientOcclusion",
-        concreteAmbientOcclusion,
-        scene,
-        progressMonitor,
-    );
-
-    // Crate
-    const crateAlbedoPromise = loadTextureAsync("CrateAlbedo", crateAlbedo, scene, progressMonitor);
-    const crateNormalPromise = loadTextureAsync("CrateNormal", crateNormal, scene, progressMonitor);
-    const crateMetallicRoughnessPromise = loadTextureAsync(
-        "CrateMetallicRoughness",
-        crateMetallicRoughness,
-        scene,
-        progressMonitor,
-    );
-    const crateAmbientOcclusionPromise = loadTextureAsync(
-        "CrateAmbientOcclusion",
-        crateAmbientOcclusion,
-        scene,
-        progressMonitor,
-    );
-
-    const treeAlbedo = await treeAlbedoPromise;
-    treeAlbedo.hasAlpha = true;
-
     // Assemble and return the textures structure
     return {
         terrains: {
@@ -323,40 +198,7 @@ export async function loadTextures(scene: Scene, progressMonitor: ILoadingProgre
             flare: await flareTexturePromise,
             butterfly: await butterflyPromise,
         },
-        materials: {
-            solarPanel: {
-                albedo: await solarPanelAlbedoPromise,
-                normal: await solarPanelNormalPromise,
-                metallicRoughness: await solarPanelMetallicRoughnessPromise,
-            },
-            spaceStation: {
-                albedo: await spaceStationAlbedoPromise,
-                normal: await spaceStationNormalPromise,
-                metallicRoughness: await spaceStationMetallicRoughnessPromise,
-                ambientOcclusion: await spaceStationAmbientOcclusionPromise,
-            },
-            metalPanels: {
-                albedo: await metalPanelsAlbedoPromise,
-                normal: await metalPanelsNormalPromise,
-                metallicRoughness: await metalPanelsMetallicRoughnessPromise,
-                ambientOcclusion: await metalPanelsAmbientOcclusionPromise,
-            },
-            concrete: {
-                albedo: await concreteAlbedoPromise,
-                normal: await concreteNormalPromise,
-                metallicRoughness: await concreteMetallicRoughnessPromise,
-                ambientOcclusion: await concreteAmbientOcclusionPromise,
-            },
-            crate: {
-                albedo: await crateAlbedoPromise,
-                normal: await crateNormalPromise,
-                metallicRoughness: await crateMetallicRoughnessPromise,
-                ambientOcclusion: await crateAmbientOcclusionPromise,
-            },
-            tree: {
-                albedo: treeAlbedo,
-            },
-        },
+        materials: await loadMaterialTextures(scene, progressMonitor),
         gasPlanet: await loadGasPlanetTextures(scene, progressMonitor),
         rings: await loadRingsTextures(scene, progressMonitor),
         environment: {
